examples: move function declarations and enums to header file for next patch
